adicionada alterações no service

+++ src/main/java/br/com/loja/projetojavaglobo/service/ProdutoService.java
@@ -1,19 +1,24 @@
 package br.com.loja.projetojavaglobo.service;
 
+import br.com.loja.projetojavaglobo.exceptions.ProductNotFoundException;
 import br.com.loja.projetojavaglobo.model.dto.RequestProdutoDto;
 import br.com.loja.projetojavaglobo.model.entities.Produto;
 import br.com.loja.projetojavaglobo.repository.ProdutoRepository;
 import com.fasterxml.jackson.databind.introspect.TypeResolutionContext;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
-@RequiredArgsConstructor
 public class ProdutoService {
     private final ProdutoRepository produtoRepository;
+    @Autowired
+    public ProdutoService(ProdutoRepository produtoRepository){
+        this.produtoRepository = produtoRepository;
+    }
     public Produto createProduto(RequestProdutoDto requestProdutoDto){
         var produto = new Produto();
         produto.setName(requestProdutoDto.getName());
@@ -25,27 +30,19 @@ public class ProdutoService {
         return savedProduto;
     }
 
-    public List<Produto> findProdutos() throws Exception {
-        List<Produto> produtoList = produtoRepository.findAll();
-
-        if (produtoList.isEmpty()) throw new Exception("Não encontramos produtos cadastrados.");
-
-        return produtoList;
+    public List<Produto> findProdutos() {
+        return produtoRepository.findAll();
     }
 
-    public Produto findById(Long id) throws Exception {
-        Optional<Produto> produtoOptional = produtoRepository.findById(id);
-
-        if (produtoOptional.isEmpty()) throw new Exception("Esse produto não está cadastrado em nossa base de dados.");
-
-        return produtoOptional.get();
+    public Produto findById(Long id) {
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException("Produto não encontrado com Id: " + id));
     }
 
-   /* public String deleteProduto(Long id) throws Exception {
-        Produto produto = produtoRepository.deleteById(id);
-
-        if (produto == null) throw new Exception("Esse produto não está cadastrado em nossa base de dados.");
-
-        return "Produto removido com sucesso.";
-    }*/
+   public void deleteProduto(Long id) {
+        if (!produtoRepository.existsById(id)){
+            throw new ProductNotFoundException("Produto não encontrado com Id: " + id);
+        }
+        produtoRepository.deleteById(id);
+    }
 }
